refactor(ccw): refactor ccw/pr/pr_management.go - large pr man...
- Code refactoring

Resolves #254
