+++ .worktree-config.json
@@ -1,10 +1,10 @@
 {
   "base_path": ".",
-  "branch_name": "issue-200-20250531-004150",
-  "worktree_path": "issue-200-20250531-004150",
-  "issue_number": 200,
-  "created_at": "2025-05-31T00:41:50.402033+09:00",
+  "branch_name": "issue-254-20250531-152647",
+  "worktree_path": "issue-254-20250531-152647",
+  "issue_number": 254,
+  "created_at": "2025-05-31T15:26:47.745803+09:00",
   "owner": "fumiya-kume",
   "repository": "FeLangKit",
-  "issue_url": "https://github.com/fumiya-kume/FeLangKit/issues/200"
+  "issue_url": "https://github.com/fumiya-kume/FeLangKit/issues/254"
 }
